🩹 Fix failing isodate test
Test failed because Node parses a string into a date using the local time zone,
unless the time zone is specifically stated. So this test passed locally, as it
was parsed in a time zone with daylight savings, but failed on the CI as it was
being probably being parsed in a time zone without daylight savings.

This is fixed by setting the time zone manually and restoring it after the tests
have finished.